urlpattern

Module quirks

This module contains functions required to integrate this library into browsers. If you are not building a browser, you can ignore this module.

Re-exports

pub use crate::Error;

Structs

EcmaRegexp 🔒

MatchInput

Matcher

UrlPattern

UrlPatternComponent

UrlPatternInit

Enums

InnerMatcher

StringOrInit

Functions

parse_match_input

parse_pattern

Parse a pattern into its components.

process_construct_pattern_input

This function constructs a UrlPattern given a string or UrlPatternInit and optionally a base url.

process_match_input

Type Aliases

Inputs