Crate urlpattern

rust-urlpattern is an implementation of the URLPattern standard for the Rust programming language.

For a usage example, see the UrlPattern documentation.

Modules

canonicalize_and_process ğŸ”’

component ğŸ”’

constructor_parser ğŸ”’

error ğŸ”’

matcher ğŸ”’

parser ğŸ”’

quirks

This module contains functions required to integrate this library into browsers. If you are not building a browser, you can ignore this module.

regexp ğŸ”’

tokenizer ğŸ”’

Structs

UrlPattern

A UrlPattern that can be matched against.

UrlPatternComponentResult

A result of a URL pattern match on a single component.

UrlPatternInit

The structured input used to create a URL pattern.

UrlPatternOptions

Options to create a URL pattern.

UrlPatternResult

A result of a URL pattern match.

Enums

Error

A error occurring during URL pattern construction, or matching.

UrlPatternMatchInput

Functions

hostname_pattern_is_ipv6_address ğŸ”’

is_absolute_pathname ğŸ”’